# perf: render list