- Get rid of unnecessary jQuery
- Construct HTML from scratch instead of parsing from string and re-querying the needed elements
- Strip HTML tags from title field only if the fieldtype is expected to contain HTML (which usually is not the case)
- Avoid useless JSON parsing (`JSON.parse("[]")`)

### Stats

Total execution time by method, rendering 100 rows (without html subject field), average from 3 tests

`get_subject_html`: 3x speedup

- Before: > 30 ms
- Now: < 10 ms

`get_meta_html`: 2x speedup

- Before: ~ 15 ms
- Now: ~ 7 ms

---

Total execution time, rendering 2500 rows, average from 3 tests:

Before: ~ 1.40 seconds
After: < 500 ms